Almost finished Mailer, suggestion in description
I created the gmail account to send the email from. Something I was
thinking about though, would be to have the user enter their email and
password. Encryption would not need to be very difficult, if required at
all because the data is only in memory. I have a working test that sends
an attached email successfully.