Updated JavaDoc, built artifacts
I reran the JavaDoc generator and updated the documentation. I also
decided to configure the application into an artifact, namely a .jar
file. It is located within /$PROJECT_DIR/out/ and should run what we
have as of this moment. I figure we should have a "nightly" build that
is working to see the progress. Compilation only takes a few seconds, so
it shouldn't be hard to build the project after major changes.